Updated code contribution readme.
This version should now be entirely up to date with the game balance /
update fixes from Wildcard's branch.

Please make an issue if I missed anything.

Planned to fix:
Not using library exists for various things.
Add RTD/Goomba integration (I plan to do some of it differently than
wildcard).

And the 3-4 updates that I haven't been able to do because of the 1.7
port.